Feat. 스플래쉬 스크린 커스텀 훅 생성
